Add files via upload
https://papers.ssrn.com/sol3/papers.cfm?abstract_id=2837598 

Open Cars
Berkeley Technology Law Journal, Vol. 32, No. 2, 2017

UC Hastings Research Paper No. 213

74 Pages Posted: 12 Sep 2016 Last revised: 8 Feb 2019
Lothar Determann
Baker & McKenzie LLP; Freie Universität Berlin; Berkeley School of Law; University of California Hastings College of the Law

Bruce Perens
Independent

Date Written: September 11, 2016

Abstract
The car of the future will be autonomous, connected and full of innovative information technology features. We may drive it or let it drive us. We know it will be a computer network on wheels. What we do not know how open the car of the future will be? Will it be like a desktop PC upon which we can select either Windows or Linux and choose a video card that meets our specific needs, or as closed as a DVD player with region control? In our article, we examine facts and arguments regarding how open the car can, should and potentially will be, as a matter of technology, economics, public policy and law. To make our points, we will tell a tale of two cars: It may be open, it may be closed. It may be the best of cars, it may be the worst of cars. We do not aim for an exact prediction or recommendation regarding the degree of openness for future cars. Rather, we intend to start or contribute to the public discussion, and contribute to the strategic planning of companies, by highlighting the economic and policy interests as well as legal rules regarding the opening or closing of automotive designs. To qualify as an open car, an automotive product must be sufficiently accessible to enable technology upgrades, aftermarket products and testing by security researchers. It must have open interfaces and openly disclosed software and hardware. It will thrive if it is associated with open developer platforms. The open car does not, however, need to run on open data, but can instead protect data privacy and security as well as or better than proprietary automotive products do today. Nor does the open car require that manufacturers install open source software. Compared to the closed, fully-proprietary car, the open car comes out ahead based on technology, competition, sustainability and environmental policy considerations. Its enemies are citing concerns regarding cybersecurity, safety and data privacy; but upon closer review, risks in these areas do not truly justify roadblocks for open cars and rather support increased openness.